split aspctlog in 2 files
